feat: Update select & input theme
